requirements.txt
# Elasticsearch
elasticsearch==8.17.0

# AI/ML
anthropic==0.77.1
google-generativeai==0.8.6

# Cloud integrations
slack-sdk==3.33.0
kubernetes==29.0.0

# Data processing
pyyaml==6.0.2
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.2

# Utilities
requests==2.32.3
aiohttp==3.11.2
rich==13.9.4
tqdm==4.67.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Code quality
black==24.10.0
pylint==3.3.2
mypy==1.13.0

# Development
ipython==8.29.0
jupyter==1.1.1
